:coffin: remove all installation scripts
In the first place, I found that having scripts like Ansible that
install everything on my computer should be a good idea... but it's not
the case.
I really dislike the fact that this script just "runs" some `apt install`
or `dnf install` without giving me any advice on what will be install
neither if I want them to be installed.